Response contentLen int -> long,

+++ nioserver/src/main/java/org/arivu/nioserver/ResponseImpl.java
@@ -24,7 +24,7 @@ final class ResponseImpl implements Response {
 
 	String redirectUrl = null;
 	
-	int contentLength = 0;
+	long contentLength = 0;
 	
 	ResponseImpl(Request request, Map<String, List<Object>> headers) {
 		this.request = request;
@@ -162,7 +162,7 @@ final class ResponseImpl implements Response {
 	 * @see org.arivu.nioserver.Response#getContentLength()
 	 */
 	@Override
-	public int getContentLength() {
+	public long getContentLength() {
 		return contentLength;
 	}
